I changed price into right requirement

+++ database/migrations/2025_07_15_154221_create_course_lists_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('course_lists', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('price', 8, 3);
+            $table->bigInteger('price'); 
             $table->integer('period')->default(12);
             $table->timestamps();
         });
